SocialServices-Refactoring: Moved entity classes to entity package in models one.
