+++ src/main/java/com/umssonline/social/controllers/CommentsRestControllerImpl.java
@@ -2,7 +2,7 @@ package com.umssonline.social.controllers;
 
 import com.umssonline.social.models.dto.CreateCommentDto;
 import com.umssonline.social.models.dto.UpdateCommentDto;
-import com.umssonline.social.models.Comment;
+import com.umssonline.social.models.entity.Comment;
 import com.umssonline.social.services.SocialService;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
