+++ src/main/java/com/umssonline/social/controllers/MessagesRestController.java
@@ -2,7 +2,7 @@ package com.umssonline.social.controllers;
 
 import com.umssonline.social.models.dto.CreateMessageDto;
 import com.umssonline.social.models.dto.UpdateMessageDto;
-import com.umssonline.social.models.Message;
+import com.umssonline.social.models.entity.Message;
 import org.springframework.http.ResponseEntity;
 
 import java.util.Collection;
