+++ src/main/java/com/umssonline/social/controllers/SharesRestController.java
@@ -1,6 +1,6 @@
 package com.umssonline.social.controllers;
 
-import com.umssonline.social.models.Share;
+import com.umssonline.social.models.entity.Share;
 import org.springframework.http.ResponseEntity;
 
 import java.util.Collection;
