+++ src/main/java/com/umssonline/social/controllers/SharesRestControllerImpl.java
@@ -1,6 +1,6 @@
 package com.umssonline.social.controllers;
 
-import com.umssonline.social.models.Share;
+import com.umssonline.social.models.entity.Share;
 import com.umssonline.social.services.SocialService;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
