--- src/main/java/com/umssonline/social/models/ResourceType.java
@@ -1,5 +1,0 @@
-package com.umssonline.social.models;
-
-public enum ResourceType {
-    PROJECT
-}
