+++ src/main/java/com/umssonline/social/models/entity/Message.java
@@ -1,4 +1,4 @@
-package com.umssonline.social.models;
+package com.umssonline.social.models.entity;
 
 import com.fasterxml.jackson.annotation.JsonBackReference;
 import lombok.Data;
