+++ src/main/java/com/umssonline/social/models/entity/ResourceType.java
@@ -1,0 +1,5 @@
+package com.umssonline.social.models.entity;
+
+public enum ResourceType {
+    PROJECT
+}
