+++ src/main/java/com/umssonline/social/repositories/api/ExtendedMessageDao.java
@@ -1,6 +1,6 @@
 package com.umssonline.social.repositories.api;
 
-import com.umssonline.social.models.Message;
+import com.umssonline.social.models.entity.Message;
 
 public interface ExtendedMessageDao extends SocialServiceDao<Message> {
