+++ src/main/java/com/umssonline/social/repositories/api/ExtendedParticipantDao.java
@@ -1,6 +1,6 @@
 package com.umssonline.social.repositories.api;
 
-import com.umssonline.social.models.Participant;
+import com.umssonline.social.models.entity.Participant;
 
 public interface ExtendedParticipantDao extends SocialServiceDao<Participant> {
 }
