+++ src/main/java/com/umssonline/social/repositories/api/ExtendedRateDao.java
@@ -1,6 +1,6 @@
 package com.umssonline.social.repositories.api;
 
-import com.umssonline.social.models.Rate;
+import com.umssonline.social.models.entity.Rate;
 
 public interface ExtendedRateDao extends SocialServiceDao<Rate> {
 }
