+++ src/main/java/com/umssonline/social/repositories/api/ExtendedResourceDao.java
@@ -1,6 +1,6 @@
 package com.umssonline.social.repositories.api;
 
-import com.umssonline.social.models.Resource;
+import com.umssonline.social.models.entity.Resource;
 
 public interface ExtendedResourceDao extends SocialServiceDao<Resource> {
 }
