+++ src/main/java/com/umssonline/social/repositories/api/ExtendedScoreDao.java
@@ -1,6 +1,6 @@
 package com.umssonline.social.repositories.api;
 
-import com.umssonline.social.models.Score;
+import com.umssonline.social.models.entity.Score;
 
 public interface ExtendedScoreDao extends SocialServiceDao<Score> {
