+++ src/main/java/com/umssonline/social/repositories/api/ExtendedShareActionDao.java
@@ -1,6 +1,6 @@
 package com.umssonline.social.repositories.api;
 
-import com.umssonline.social.models.ShareAction;
+import com.umssonline.social.models.entity.ShareAction;
 
 public interface ExtendedShareActionDao extends SocialServiceDao<ShareAction> {
 }
