+++ src/main/java/com/umssonline/social/repositories/api/ExtendedShareDao.java
@@ -1,6 +1,6 @@
 package com.umssonline.social.repositories.api;
 
-import com.umssonline.social.models.Share;
+import com.umssonline.social.models.entity.Share;
 
 public interface ExtendedShareDao extends SocialServiceDao<Share> {
