+++ src/main/java/com/umssonline/social/services/ParticipantServiceImpl.java
@@ -1,6 +1,6 @@
 package com.umssonline.social.services;
 
-import com.umssonline.social.models.Participant;
+import com.umssonline.social.models.entity.Participant;
 import com.umssonline.social.repositories.api.ExtendedParticipantDao;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
