+++ src/main/java/com/umssonline/social/services/ResourceServiceImpl.java
@@ -1,6 +1,6 @@
 package com.umssonline.social.services;
 
-import com.umssonline.social.models.Resource;
+import com.umssonline.social.models.entity.Resource;
 import com.umssonline.social.repositories.api.ExtendedResourceDao;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
