+++ src/main/java/com/umssonline/social/services/ShareServiceImpl.java
@@ -1,8 +1,8 @@
 package com.umssonline.social.services;
 
-import com.umssonline.social.models.Participant;
-import com.umssonline.social.models.Resource;
-import com.umssonline.social.models.Share;
+import com.umssonline.social.models.entity.Participant;
+import com.umssonline.social.models.entity.Resource;
+import com.umssonline.social.models.entity.Share;
 import com.umssonline.social.repositories.api.ExtendedShareDao;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
